+++ guillotina_elasticsearch/commands/fields.py
@@ -8,7 +8,7 @@ import operator
 
 
 class FieldsCommand(Command):
-    description = 'Report on configured fields'
+    description = "Report on configured fields"
     type_counts = {}
     schema_counts = {}
     total = stored = 0
@@ -17,50 +17,53 @@ class FieldsCommand(Command):
 
     def get_parser(self):
         parser = super().get_parser()
-        parser.add_argument('--summary', action='store_true')
-        parser.add_argument('--schema', action='append')
-        parser.add_argument('--type', action='append')
+        parser.add_argument("--summary", action="store_true")
+        parser.add_argument("--schema", action="append")
+        parser.add_argument("--type", action="append")
         return parser
 
     def _count_field(self, field, schemas=None):
-        if 'properties' in field:
-            for sub_field in field['properties'].values():
-                self._count_field(sub_field, field['_schemas'])
+        if "properties" in field:
+            for sub_field in field["properties"].values():
+                self._count_field(sub_field, field["_schemas"])
             return
 
         if schemas is None:
-            schemas = field['_schemas']
+            schemas = field["_schemas"]
 
         self.total += 1
-        if field.get('store'):
+        if field.get("store"):
             self.stored += 1
-        if field['type'] not in self.type_counts:
-            self.type_counts[field['type']] = 0
+        if field["type"] not in self.type_counts:
+            self.type_counts[field["type"]] = 0
         for schema_name in schemas:
             if schema_name not in self.schema_counts:
                 self.schema_counts[schema_name] = 0
             self.schema_counts[schema_name] += 1
-        self.type_counts[field['type']] += 1
+        self.type_counts[field["type"]] += 1
 
     def summary(self):
-        for field in get_mappings(self.selected_schemas, schema_info=True)['properties'].values():  # noqa
+        for field in get_mappings(self.selected_schemas, schema_info=True)[
+            "properties"
+        ].values():  # noqa
             self._count_field(field)
 
-        pprint({
-            'total': self.total,
-            'stored': self.stored,
-            'type_counts': sorted(
-                self.type_counts.items(), key=operator.itemgetter(1),
-                reverse=True),
-            'schema_counts': sorted(
-                self.schema_counts.items(), key=operator.itemgetter(1),
-                reverse=True)
-        })
+        pprint(
+            {
+                "total": self.total,
+                "stored": self.stored,
+                "type_counts": sorted(
+                    self.type_counts.items(), key=operator.itemgetter(1), reverse=True
+                ),
+                "schema_counts": sorted(
+                    self.schema_counts.items(), key=operator.itemgetter(1), reverse=True
+                ),
+            }
+        )
 
     async def run(self, arguments, settings, app):
         if arguments.schema:
-            self.selected_schemas = [
-                resolve_dotted_name(s) for s in arguments.schema]
+            self.selected_schemas = [resolve_dotted_name(s) for s in arguments.schema]
         if arguments.type:
             if self.selected_schemas is None:
                 self.selected_schemas = []
@@ -70,6 +73,5 @@ class FieldsCommand(Command):
         if self.arguments.summary:
             self.summary()
         else:
-            fields = get_mappings(
-                self.selected_schemas, schema_info=True)['properties']
+            fields = get_mappings(self.selected_schemas, schema_info=True)["properties"]
             pprint(fields)